+++ src/main/kotlin/com/whalesea/ideatabmanager/toolwindow/TabGroupIcons.kt
@@ -9,4 +9,5 @@ object TabGroupIcons {
     val expand: Icon = IconLoader.getIcon("/icons/expandGroup.svg", TabGroupIcons::class.java)
     val newEmptyGroup: Icon = IconLoader.getIcon("/icons/newEmptyGroup.svg", TabGroupIcons::class.java)
     val saveCurrentTabs: Icon = IconLoader.getIcon("/icons/saveCurrentTabs.svg", TabGroupIcons::class.java)
+    val saveSelectedTabs: Icon = IconLoader.getIcon("/icons/saveSelectedTabs.svg", TabGroupIcons::class.java)
 }